requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pydantic==2.12.5
requests==2.32.5
schedule==1.2.2
rich==14.3.3
python-multipart==0.0.20
python-dotenv==1.2.1
